Migrate `@zeit/next-bundle-analyzer` to `@next/bundle-analyzer` (#20)
* Migrate @zeit/next-bundle-analyzer to @next/bundle-analyzer

In `E2-lazy-loading-modules`, it uses `@next/bundle-analyzer`, whereas in `E3-lazy-loading-components`, it uses `@zeit/next-bundle-analyzer`. That causes a confusion which library should be used.

I checked in https://www.npmjs.com/package/@zeit/next-bundle-analyzer and `@zeit/next-bundle-analyzer` was last publish a year ago, but `@next/bundle-analyzer` was last publish 3 days ago (https://www.npmjs.com/package/@next/bundle-analyzer). This makes me come to a conclusion that `@zeit/next-bundle-analyzer` was deprecated in favor of `@next/bundle-analyzer`, and we should migrate to the new one.

* Update scripts